payment_service/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.31.0 